+++ src/mibig_input_script/classes/class_base.py
@@ -119,6 +119,7 @@ class BaseClass:
     const_pubmed_id_regexp = r"\d+"
     const_url_regexp = r"https?:\/\/(www\.)?[-a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)"
     const_ncbi_accession_regexp = r"^([A-Za-z0-9_]{3,}\.\d)|(MIBIG\.BGC\d{7}\.\d)$"
+    const_proteinogenic_aa_regexp = r"^[A-Z]+$"
 
     const_ncbi_acc_illegal_chars = [
         ",",