+++ src/mibig_input_script/classes/class_mibig_entry.py
@@ -258,9 +258,9 @@ class MibigEntry(BaseClass):
         input_message = [
             (
                 "================================================\n"
-                "Enter the biosynthetic class of the BGC.\n",
+                "Enter the biosynthetic class of the BGC.\n"
                 "Separate multiple entries with a TAB character.\n"
-                "================================================\n",
+                "================================================\n"
             )
         ]
         for option in self.const_biosyn_classes:
@@ -361,8 +361,6 @@ class MibigEntry(BaseClass):
 
         Returns:
             None
-
-        Notes:
         """
         input_msg_number = (
             "================================================\n"