+++ src/mibig_input_script/classes/class_ripp.py
@@ -1,11 +1,10 @@
 """
 Module to get information about RiPP.
 
-Module to get and store optional RiPP information. Not fully implemented
-yet.
+Module to get and store optional RiPP information.
 """
 from copy import deepcopy
-from typing import Dict, Self
+from typing import Dict, Self, List
 import re
 
 from mibig_input_script.classes.class_base import BaseClass
@@ -18,18 +17,32 @@ class Ripp(BaseClass):
         mibig_dict (Dict) : holding the existing mibig entry
 
     Methods:
-        initialize_ripp_entry(self: Self) -> None
         export_dict(self: Self) -> Dict
+        initialize_ripp_entry(self: Self) -> None
         set_flag_minimal_false(self: Self) -> None
         get_input(self: Self) -> None
         get_cyclic(self: Self) -> None
         get_subclass(self: Self) -> None
-
-
-
-        get_mod_precursor_pep(self: Self) -> None
-        get_precursor_peptide_entry(self: Self, index: int | str) -> None
-
+        get_peptidases(self: Self) -> None
+        get_precursor_genes(self: Self) -> None
+        remove_precursor_gene_entry(self: Self, invar_menu_entries: int) -> None
+        get_precursor_gene_entry(self: Self, index: int | str) -> None
+        write_precursor_gene_entry(
+            self: Self,
+            index : int,
+            input_leader : str,
+            input_core : List,
+            input_follower : str,
+            input_gene_id : str,
+            input_cleavage : List,
+            input_recognition : str,
+        ) -> None
+        get_precursor_gene_leader(self: Self) -> str | bool | None
+        get_precursor_gene_core(self: Self) -> List | bool
+        get_precursor_gene_follower(self: Self) -> str | bool | None
+        get_precursor_gene_id(self: Self) -> str | bool
+        get_precursor_gene_cleavage(self: Self) -> List | bool | None
+        get_precursor_recognition(self: Self) -> str | bool | None
     """
 
     def __init__(self: Self, mibig_entry: Dict):
@@ -92,6 +105,7 @@ class Ripp(BaseClass):
             "1": self.get_cyclic,
             "2": self.get_subclass,
             "3": self.get_peptidases,
+            "4": self.get_precursor_genes,
         }
 
         while True:
@@ -110,6 +124,13 @@ class Ripp(BaseClass):
             except KeyError:
                 peptidases = "None"
 
+            try:
+                precursor_genes = len(
+                    self.mibig_dict["cluster"]["ripp"]["precursor_genes"]
+                )
+            except KeyError:
+                precursor_genes = "None"
+
             input_message = (
                 "================================================\n"
                 "RiPP annotation (currently not fully implemented):\n"
@@ -120,6 +141,7 @@ class Ripp(BaseClass):
                 f"1) Cyclic (currently: '{cyclic}')\n"
                 f"2) RiPP subclass (currently: '{subclass}')\n"
                 f"3) RiPP-cleaving peptidases (currently: '{peptidases}')\n"
+                f"4) RiPP precursor genes (currently: '{precursor_genes}')\n"
                 "================================================\n"
             )
 
@@ -178,13 +200,13 @@ class Ripp(BaseClass):
         """
         input_msg_subclass = (
             "================================================\n"
-            "Enter the subclass of the RiPP (a single one).\n"
+            "Enter the subclass of the RiPP (or SKIP by pressing enter).\n"
             "================================================\n"
         )
 
         input_subclass = input(input_msg_subclass)
         if input_subclass == "":
-            self.error_message_formatted("Invalid input provided")
+            self.message_formatted("Invalid input provided - SKIP")
             return
         else:
             self.mibig_dict["cluster"]["ripp"]["subclass"] = input_subclass
@@ -202,6 +224,7 @@ class Ripp(BaseClass):
         input_msg_peptidases = (
             "================================================\n"
             "Enter the peptidase(s) involved in precursor cleavage.\n"
+            "To SKIP, press enter.\n"
             "To specify multiple entries, separate them with a TAB character.\n"
             "================================================\n"
         )
@@ -209,7 +232,7 @@ class Ripp(BaseClass):
         input_peptidases = list(filter(None, input_peptidases.split("\t")))
 
         if len(input_peptidases) == 0:
-            self.error_message_formatted("Empty input value")
+            self.message_formatted("Empty input value - SKIP")
             return
         else:
             peptidases_set = set()
@@ -218,14 +241,8 @@ class Ripp(BaseClass):
             self.mibig_dict["cluster"]["ripp"]["peptidases"] = list(peptidases_set)
             return
 
-    # get a list as input
-    # assign to a dict
-    # return
-    # add to selection screen
-
-    ####
-    def get_mod_precursor_pep(self: Self) -> None:
-        """Get input on the modified precursor peptide structure.
+    def get_precursor_genes(self: Self) -> None:
+        """Get input on precursor genes.
 
         Parameters:
             `self` : The instance of class RiPP.
@@ -234,137 +251,439 @@ class Ripp(BaseClass):
             None
         """
         try:
-            self.mibig_dict["cluster"]["ripp"]["mod_precursor_pep"]
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"]
         except KeyError:
-            self.mibig_dict["cluster"]["ripp"]["mod_precursor_pep"] = []
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"] = []
+
+        invar_menu_entries = 3
 
         while True:
-            input_message_list = [
+            input_msg_menu = [
                 (
                     "================================================\n"
-                    "Add new modified precursor peptides or modify existing ones:\n"
+                    "Add new precursor genes or modify existing ones:\n"
                     "Enter a number and press enter.\n"
                     "================================================\n"
                     "0) Save and continue\n"
                     "1) Add new entry\n"
+                    "2) Delete an entry\n"
                 )
             ]
 
-            counter = 2
+            counter = 3
             for entry in range(
-                len(self.mibig_dict["cluster"]["ripp"]["mod_precursor_pep"])
+                len(self.mibig_dict["cluster"]["ripp"]["precursor_genes"])
             ):
-                input_message_list.append(
+                input_msg_menu.append(
                     (
-                        f"{counter}) "
-                        f"{self.mibig_dict['cluster']['ripp']['mod_precursor_pep'][entry]['name']}"
+                        f"{counter}) Gene ID: "
+                        f"{self.mibig_dict['cluster']['ripp']['precursor_genes'][entry]['gene_id']}"
                         f"\n"
                     )
                 )
                 counter += 1
 
-            input_message_list.append(
+            input_msg_menu.append(
                 ("================================================\n")
             )
-            input_message = "".join([i for i in input_message_list])
+            input_msg_menu = "".join([i for i in input_msg_menu])
 
-            input_raw = input(input_message)
+            input_selection = input(input_msg_menu)
 
-            if input_raw == "":
+            if input_selection == "":
                 self.error_message_formatted("Invalid input provided")
                 continue
-            elif input_raw == "0":
+            elif input_selection == "0":
                 break
-            elif input_raw == "1":
-                self.get_precursor_peptide_entry(input_raw)
+            elif input_selection == "1":
+                self.get_precursor_gene_entry(input_selection)
+                continue
+            elif input_selection == "2":
+                self.remove_precursor_gene_entry(invar_menu_entries)
                 continue
             else:
                 try:
-                    input_raw = int(input_raw) - 2
+                    input_selection = int(input_selection) - invar_menu_entries
                 except ValueError:
                     self.error_message_formatted("Invalid input provided")
                     continue
-                if input_raw <= len(
-                    self.mibig_dict["cluster"]["ripp"]["mod_precursor_pep"]
+                if (
+                    0
+                    <= input_selection
+                    <= len(self.mibig_dict["cluster"]["ripp"]["precursor_genes"])
                 ):
-                    self.get_precursor_peptide_entry(input_raw)
+                    self.get_precursor_gene_entry(input_selection)
                     continue
                 else:
                     self.error_message_formatted("Invalid input provided")
                     continue
+
         return
 
-    def get_precursor_peptide_entry(self: Self, index: int | str) -> None:
-        """Get a new modified precursor peptide
+    def remove_precursor_gene_entry(self: Self, invar_menu_entries: int) -> None:
+        """Remove a precursor gene entry from the MIBiG entry.
 
         Parameters:
-            `self` : The instance of class RiPP.
+            `self` : The instance of class MibigEntry.
 
         Returns:
-            `str` to create a new entry (append), `int` to replace existing
+            None
+        """
+        input_msg_number = (
+            "================================================\n"
+            "Enter the number of precursor gene to remove (a single one):\n"
+            "================================================\n"
+        )
+        input_number = input(input_msg_number)
+
+        try:
+            input_number = int(input_number)
+        except ValueError:
+            self.error_message_formatted("Invalid input value")
+            return
+
+        input_number = input_number - invar_menu_entries
+
+        if (
+            0
+            <= input_number
+            <= len(self.mibig_dict["cluster"]["ripp"]["precursor_genes"])
+        ):
+            if self.ask_proceed_input():
+                self.mibig_dict["cluster"]["ripp"]["precursor_genes"].pop(input_number)
+                return
+            else:
+                return
+        else:
+            self.error_message_formatted("Entry not found")
+            return
+
+    def get_precursor_gene_entry(self: Self, index: int | str) -> None:
+        """Add a precursor gene entry.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+            `str` to create a new entry (append), `int` to replace existing.
+
+        Returns:
+            None
 
         Notes:
             If `index` is `str`, then a new entry is written (appended).
             If `index` is `int`, an existing entry is modified (overwritten).
         """
-        input_msg_gene_id = (
-            "================================================\n"
-            "Enter the gene ID:\n"
-            "================================================\n"
-        )
-        input_msg_peptide = (
-            "================================================\n"
-            "Enter the modified precursor peptide using the shorthand format:\n"
-            "(leader)-core-(follower) (e.g. MR-YYH-)\n"
-            "================================================\n"
-        )
-        input_msg_name = (
-            "================================================\n"
-            "Enter the name of the final product:\n"
-            "================================================\n"
-        )
-
-        input_gene_id = input(input_msg_gene_id)
-
-        if input_gene_id == "":
-            self.error_message_formatted("Invalid input provided")
-            return
-        elif not re.match(r"^[^, ]*$", input_gene_id):
-            self.error_message_formatted("Input value cannot contain space or comma")
+        input_leader = self.get_precursor_gene_leader()
+        if input_leader == False:
             return
         else:
             pass
 
-        input_peptide = input(input_msg_peptide)
-
-        if input_peptide == "":
-            self.error_message_formatted("Invalid input provided")
+        input_core = self.get_precursor_gene_core()
+        if input_core == False:
             return
-        elif not re.match(r"^[A-Z]*-.+-[A-Z]*$", input_peptide):
-            self.error_message_formatted("Modified precursor peptide in wrong format")
+        elif input_core is None:
+            self.error_message_formatted("Core peptide cannot be empty")
             return
         else:
             pass
 
-        input_name = input(input_msg_name)
+        input_follower = self.get_precursor_gene_follower()
+        if input_follower == False:
+            return
+        else:
+            pass
 
-        if input_peptide == "":
-            self.error_message_formatted("Invalid input provided")
+        input_gene_id = self.get_precursor_gene_id()
+        if input_gene_id == False:
+            return
+        elif input_gene_id is None:
+            self.error_message_formatted("Gene ID cannot be empty")
+            return
+        else:
+            pass
+
+        input_cleavage = self.get_precursor_gene_cleavage()
+        if input_cleavage == False:
+            return
+        else:
+            pass
+
+        input_recognition = self.get_precursor_recognition()
+        if input_recognition == False:
             return
         else:
             pass
 
         if isinstance(index, str):
-            self.mibig_dict["cluster"]["ripp"]["mod_precursor_pep"].append(
-                {"gene_id": input_gene_id, "peptide": input_peptide, "name": input_name}
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"].append({})
+            index = -1
+            self.write_precursor_gene_entry(
+                index,
+                input_leader,
+                input_core,
+                input_follower,
+                input_gene_id,
+                input_cleavage,
+                input_recognition,
             )
         elif isinstance(index, int):
-            self.mibig_dict["cluster"]["ripp"]["mod_precursor_pep"][index] = {
-                "gene_id": input_gene_id,
-                "peptide": input_peptide,
-                "name": input_name,
-            }
+            self.write_precursor_gene_entry(
+                index,
+                input_leader,
+                input_core,
+                input_follower,
+                input_gene_id,
+                input_cleavage,
+                input_recognition,
+            )
         else:
             pass
 
         return
+
+    def write_precursor_gene_entry(
+        self: Self,
+        index: int,
+        input_leader: str,
+        input_core: List,
+        input_follower: str,
+        input_gene_id: str,
+        input_cleavage: List,
+        input_recognition: str,
+    ) -> None:
+        """Writes precursor gene entry data to self.mibig_dict
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+            `index` : The index int of precursor gene to write data to
+            `input_leader` : A str of the leader peptide sequence
+            `input_core` : A List of core sequences
+            `input_follower` : A str of the follower peptide sequence
+            `input_gene_id` : A str of the gene ID of the precursor gene
+            `input_cleavage` : A List of cleavage recognition sites
+            `input_recognition` : A str of the enzyme recognition motif
+
+        Returns:
+            None
+        """
+        if input_leader is not None:
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"][index][
+                "leader_sequence"
+            ] = input_leader
+        else:
+            pass
+
+        if input_core is not None:
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"][index][
+                "core_sequence"
+            ] = input_core
+        else:
+            pass
+
+        if input_follower is not None:
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"][index][
+                "follower_sequence"
+            ] = input_follower
+        else:
+            pass
+
+        if input_gene_id is not None:
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"][index][
+                "gene_id"
+            ] = input_gene_id
+        else:
+            pass
+
+        if input_cleavage is not None:
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"][index][
+                "cleavage_recogn_site"
+            ] = input_cleavage
+        else:
+            pass
+
+        if input_recognition is not None:
+            self.mibig_dict["cluster"]["ripp"]["precursor_genes"][index][
+                "recognition_motif"
+            ] = input_recognition
+        else:
+            pass
+
+        return
+
+    def get_precursor_gene_leader(self: Self) -> str | bool | None:
+        """Get the precursor gene leader sequence.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+
+        Returns:
+            Return a str of the precursor gene leader sequence or
+            bool (error) or None
+        """
+        input_msg_leader = (
+            "================================================\n"
+            "Enter the leader peptide sequence (or SKIP by pressing enter):\n"
+            "Only enter proteinogenic amino acids (upper-case letters).\n"
+            "================================================\n"
+        )
+        input_leader = input(input_msg_leader)
+        if input_leader == "":
+            self.message_formatted("Empty input value - SKIP")
+            return None
+        elif not re.match(self.const_proteinogenic_aa_regexp, input_leader):
+            self.error_message_formatted(
+                f"'{input_leader}' is not in single-letter proteinogenic amino acid code"
+            )
+            return False
+        else:
+            return input_leader
+
+    def get_precursor_gene_core(self: Self) -> List | bool:
+        """Get a list of precursor gene core sequences.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+
+        Returns:
+            Return the list of precursor gene core sequences or bool (error)
+        """
+        input_msg_core = (
+            "================================================\n"
+            "Enter the core peptide sequence(s):\n"
+            "Only enter proteinogenic amino acids (upper-case letters).\n"
+            "To specify multiple ones, separate entries with a TAB character.\n"
+            "================================================\n"
+        )
+        input_core = input(input_msg_core)
+        input_core = list(filter(None, input_core.split("\t")))
+        if len(input_core) == 0:
+            self.error_message_formatted("Core peptide sequences cannot be empty")
+            return False
+        else:
+            input_core_set = set()
+            for entry in input_core:
+                if not re.match(self.const_proteinogenic_aa_regexp, entry):
+                    self.error_message_formatted(
+                        f"'{entry}' is not in single-letter proteinogenic amino acid code"
+                    )
+                    return False
+                else:
+                    input_core_set.add(entry)
+            return list(input_core_set)
+
+    def get_precursor_gene_follower(self: Self) -> str | bool | None:
+        """Get the precursor gene follower sequence.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+
+        Returns:
+            Return a str of the precursor gene follower sequence or
+            bool (error) or None (skip)
+        """
+        input_msg_follower = (
+            "================================================\n"
+            "Enter the follower peptide sequence (or SKIP by pressing enter):\n"
+            "Only enter proteinogenic amino acids (upper-case letters).\n"
+            "================================================\n"
+        )
+        input_follower = input(input_msg_follower)
+        if input_follower == "":
+            self.message_formatted("Empty input value - SKIP")
+            return None
+        elif not re.match(self.const_proteinogenic_aa_regexp, input_follower):
+            self.error_message_formatted(
+                f"'{input_follower}' is not in single-letter proteinogenic amino acid code"
+            )
+            return False
+        else:
+            return input_follower
+
+    def get_precursor_gene_id(self: Self) -> str | bool:
+        """Get the precursor gene follower sequence.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+
+        Returns:
+            Return a str of the precursor gene id or
+            bool (error)
+        """
+        input_msg_gene_id = (
+            "================================================\n"
+            "Enter the precursor gene ID:\n"
+            "Only enter proteinogenic amino acids (upper-case letters).\n"
+            "================================================\n"
+        )
+        input_precursor_gene_id = input(input_msg_gene_id)
+        if input_precursor_gene_id == "":
+            self.error_message_formatted("The precursor gene ID cannot be empty!")
+            return False
+        elif not re.match(r"^[^, ]*$", input_precursor_gene_id):
+            self.error_message_formatted(
+                f"'{input_precursor_gene_id}' has invalid format"
+            )
+            return False
+        else:
+            return input_precursor_gene_id
+
+    def get_precursor_gene_cleavage(self: Self) -> List | bool | None:
+        """Get a list of cleavage recognition sites.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+
+        Returns:
+            Return the list of cleavage recognition sites or bool (error) or None (skip)
+        """
+        input_msg_cleavage = (
+            "================================================\n"
+            "Enter cleavage recognition site(s) (to SKIP, press enter):\n"
+            "Only enter proteinogenic amino acids (upper-case letters).\n"
+            "To specify multiple ones, separate entries with a TAB character.\n"
+            "================================================\n"
+        )
+        input_cleavage = input(input_msg_cleavage)
+        input_cleavage = list(filter(None, input_cleavage.split("\t")))
+        if len(input_cleavage) == 0:
+            self.message_formatted("Empty input value - SKIP")
+            return None
+        else:
+            input_cleavage_set = set()
+            for entry in input_cleavage:
+                if not re.match(self.const_proteinogenic_aa_regexp, entry):
+                    self.error_message_formatted(
+                        f"'{entry}' is not in single-letter proteinogenic amino acid code"
+                    )
+                    return False
+                else:
+                    input_cleavage_set.add(entry)
+            return list(input_cleavage_set)
+
+    def get_precursor_recognition(self: Self) -> str | bool | None:
+        """Get the precursor gene recognition sequence.
+
+        Parameters:
+            `self` : The instance of class MibigEntry.
+
+        Returns:
+            Return a str of the precursor gene recognition sequence or
+            bool (error) or None (skip)
+        """
+        input_msg_recognition = (
+            "================================================\n"
+            "Enter the recognition motif in the leader peptide (or SKIP by pressing enter):\n"
+            "Only enter proteinogenic amino acids (upper-case letters).\n"
+            "================================================\n"
+        )
+        input_recognition = input(input_msg_recognition)
+        if input_recognition == "":
+            self.message_formatted("Empty input value - SKIP")
+            return None
+        elif not re.match(self.const_proteinogenic_aa_regexp, input_recognition):
+            self.error_message_formatted(
+                f"'{input_recognition}' is not in single-letter proteinogenic amino acid code"
+            )
+            return False
+        else:
+            return input_recognition